Simplify if condition.
Refs: #10441

+++ features/step_definitions/apt.rb
@@ -30,9 +30,7 @@ end
 
 When /^I update APT using Synaptic$/ do
   recovery_proc = Proc.new do
-    if $vm.has_process?("synaptic")
-      step 'I kill the process "synaptic"'
-    end
+    step 'I kill the process "synaptic"' if $vm.has_process?("synaptic")
     step "I start Synaptic"
   end
   retry_tor(recovery_proc) do
@@ -56,9 +54,7 @@ end
 Then /^I should be able to install a package using Synaptic$/ do
   package = "cowsay"
   recovery_proc = Proc.new do
-    if $vm.has_process?("synaptic")
-      step 'I kill the process "synaptic"'
-    end
+    step 'I kill the process "synaptic"' if $vm.has_process?("synaptic")
     $vm.execute("apt-get -y purge #{package}")
     step "I start Synaptic"
   end
